requirements_mysql.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
Werkzeug==2.3.6
PyMySQL==1.1.0
cryptography==41.0.3
reportlab==4.0.4
